requirements.txt
Pillow~=9.1.1
mouse~=0.7.1
pynput~=1.7.6